requirements.txt
flask
pytesseract
pdf2image
pillow
opencv-python
numpy
fpdf
pikepdf

